Avoid pushing base layers
When we detect a layer we pushed before, stub in an empty layer, with an
annotation containing the uncompressed layer DiffID it's replacing. On
pull, swap in the uncompressed layer digest so the actual layer in the
layer store will be used without trying to pull/use the empty layer.